refactor: web/dto패키지 도메인별 분리
