+++ src/main/java/roomescape/service/ThemeService.java
@@ -11,8 +11,8 @@ import roomescape.domain.policy.RankingPolicy;
 import roomescape.domain.repository.ThemeRepository;
 import roomescape.exception.theme.NotFoundThemeException;
 import roomescape.exception.theme.ReservationReferencedThemeException;
-import roomescape.web.dto.request.ThemeRequest;
-import roomescape.web.dto.response.ThemeResponse;
+import roomescape.web.dto.request.theme.ThemeRequest;
+import roomescape.web.dto.response.theme.ThemeResponse;
 
 @Service
 public class ThemeService {
