+++ src/main/java/roomescape/web/controller/MemberController.java
@@ -16,8 +16,8 @@ import org.springframework.web.bind.annotation.RestController;
 
 import lombok.RequiredArgsConstructor;
 import roomescape.service.MemberService;
-import roomescape.web.dto.request.SignupRequest;
-import roomescape.web.dto.response.MemberResponse;
+import roomescape.web.dto.request.member.SignupRequest;
+import roomescape.web.dto.response.member.MemberResponse;
 
 @RestController
 @RequestMapping("/members")
