+++ src/main/java/roomescape/web/controller/ReservationTimeController.java
@@ -17,9 +17,9 @@ import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 
 import roomescape.service.ReservationTimeService;
-import roomescape.web.dto.request.ReservationTimeRequest;
-import roomescape.web.dto.response.AvailableReservationTimeResponse;
-import roomescape.web.dto.response.ReservationTimeResponse;
+import roomescape.web.dto.request.time.ReservationTimeRequest;
+import roomescape.web.dto.response.time.AvailableReservationTimeResponse;
+import roomescape.web.dto.response.time.ReservationTimeResponse;
 
 @RequestMapping("/times")
 @RestController
