+++ src/main/java/roomescape/web/controller/ThemeController.java
@@ -16,8 +16,8 @@ import org.springframework.web.bind.annotation.RestController;
 
 import roomescape.domain.policy.WeeklyRankingPolicy;
 import roomescape.service.ThemeService;
-import roomescape.web.dto.request.ThemeRequest;
-import roomescape.web.dto.response.ThemeResponse;
+import roomescape.web.dto.request.theme.ThemeRequest;
+import roomescape.web.dto.response.theme.ThemeResponse;
 
 @RequestMapping("/themes")
 @RestController
