--- src/main/java/roomescape/web/dto/request/MemberInfo.java
@@ -1,4 +1,0 @@
-package roomescape.web.dto.request;
-
-public record MemberInfo(Long id) {
-}
