+++ src/main/java/roomescape/web/dto/request/member/MemberInfo.java
@@ -1,0 +1,4 @@
+package roomescape.web.dto.request.member;
+
+public record MemberInfo(Long id) {
+}
