+++ src/main/java/roomescape/web/dto/request/time/ReservationTimeRequest.java
@@ -1,19 +1,19 @@
-package roomescape.web.dto.response;
+package roomescape.web.dto.request.time;
 
 import java.time.LocalTime;
+
+import jakarta.validation.constraints.NotNull;
 
 import com.fasterxml.jackson.annotation.JsonFormat;
 import com.fasterxml.jackson.annotation.JsonFormat.Shape;
 
 import roomescape.domain.ReservationTime;
 
-public record ReservationTimeResponse(Long id, LocalTime startAt) {
+public record ReservationTimeRequest(
+        @NotNull(message = "시간은 빈값을 허용하지 않습니다.") LocalTime startAt) {
 
-    public static ReservationTimeResponse from(ReservationTime time) {
-        return new ReservationTimeResponse(
-                time.getId(),
-                time.getStartAt()
-        );
+    public ReservationTime toReservationTime() {
+        return new ReservationTime(startAt);
     }
 
     @Override
