+++ src/main/java/roomescape/web/dto/response/member/MemberResponse.java
@@ -1,4 +1,4 @@
-package roomescape.web.dto.response;
+package roomescape.web.dto.response.member;
 
 public record MemberResponse(Long id, String name) {
