+++ src/main/java/roomescape/web/dto/response/theme/ThemeResponse.java
@@ -1,4 +1,4 @@
-package roomescape.web.dto.response;
+package roomescape.web.dto.response.theme;
 
 import roomescape.domain.Theme;
