+++ src/test/java/roomescape/acceptance/MissionStepTest.java
@@ -21,7 +21,7 @@ import roomescape.domain.Member;
 import roomescape.domain.Role;
 import roomescape.service.security.JwtUtils;
 import roomescape.web.controller.ReservationController;
-import roomescape.web.dto.response.ReservationResponse;
+import roomescape.web.dto.response.reservation.ReservationResponse;
 
 class MissionStepTest extends AcceptanceTest {
     private static final String TOKEN = JwtUtils.encode(new Member(1L, "a", "B", "c", Role.ADMIN));
