+++ src/test/java/roomescape/web/controller/AuthControllerTest.java
@@ -25,8 +25,8 @@ import roomescape.domain.Member;
 import roomescape.domain.Role;
 import roomescape.service.MemberService;
 import roomescape.service.security.JwtUtils;
-import roomescape.web.dto.request.LoginRequest;
-import roomescape.web.dto.response.MemberResponse;
+import roomescape.web.dto.request.member.LoginRequest;
+import roomescape.web.dto.response.member.MemberResponse;
 
 
 @WebMvcTest(controllers = AuthController.class)
